requirements.txt
streamlit
google-generativeai
jinja2
regex
ipaddress
